requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cos_python_sdk_v5==1.9.41
crcmod==1.7
cryptography==46.0.5
dashscope==1.25.13
distro==1.9.0
fastapi==0.135.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
lxml==6.0.2
MarkupSafe==3.0.3
multidict==6.7.1
openai==2.26.0
propcache==0.4.1
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.22
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
websocket-client==1.9.0
xmltodict==1.0.4
yarl==1.23.0
